Refactor : CHANGE MESSAGE.CLASS AND PRINT.CLASS TO FILE VIEW
