+++ src/main/java/racingcar/View/Print.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.View;
 
 import racingcar.Model.Car;
